Fix tests in AutoBufferTest
